:truck: Rename update client profile component names
From cards to forms
